Minor nits detected in final testing of 3.0. real 3.0 version
svn:revision:3061
cvs:account:mike
cvs:branch:trunk

+++ setup.sh
@@ -62,7 +62,10 @@ SCRIPTS="machinetype.sh cakeinclude.sh cray.sh"
 
 for i in ${SCRIPTS}
 do
-	mv -f ${BINDIR}/${i} ${BINDIR}/`basename ${i} .sh`.bak
+	if test -f ${BINDIR}/${i}
+	then
+		mv -f ${BINDIR}/${i} ${BINDIR}/`basename ${i} .sh`.bak
+	fi
 	cp ${i} ${BINDIR}/.
 done
